Refactor the stats out into a separate module
